pyproject.toml
[project]
name = "soul-verse-api"
version = "0.1.0"
description = ""
authors = [
    {name = "sylvaintiassou",email = "tiassousylvain6120@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "redis (>=7.1.0,<8.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "firebase-admin (>=7.1.0,<8.0.0)",
    "apscheduler (>=3.11.2,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]

[tool.poetry]
packages = [{include = "soul_verse_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
dev = "scripts.dev:dev"
prod = "scripts.prod:prod"